+++ src/roiextractors/extractors/bioformatsimagingextractor/bioformatsimagingextractor.py
